+++ content/src/controller/get-audit-handler.ts
@@ -1,0 +1,42 @@
+import { HandlerContextWithPath, InvalidRequestError, NotFoundError, parseEntityType } from '../types'
+import { GetEntityAuditInformation200 } from '@dcl/catalyst-api-specs/lib/client/client.schemas'
+import { getDeployments } from '../logic/deployments'
+import { DeploymentField } from './Controller'
+import { AuditInfo } from 'src/deployment-types'
+
+// Method: GET
+export async function getEntityAuditInformation(
+  context: HandlerContextWithPath<'database' | 'denylist' | 'metrics', '/audit/:type/:entityId'>
+): Promise<{ status: 200; body: GetEntityAuditInformation200 }> {
+  const type = parseEntityType(context.params.type)
+  const entityId = context.params.entityId
+
+  // Validate type is valid
+  if (!type) {
+    throw new InvalidRequestError(`Unrecognized type: ${context.params.type}`)
+  }
+
+  const { deployments } = await getDeployments(context.components, context.components.database, {
+    fields: [DeploymentField.AUDIT_INFO],
+    filters: { entityIds: [entityId], entityTypes: [type] },
+    includeDenylisted: true
+  })
+
+  if (deployments.length === 0) {
+    throw new NotFoundError('No deployment found')
+  }
+
+  const { auditInfo } = deployments[0]
+  const body: AuditInfo = {
+    version: auditInfo.version,
+    localTimestamp: auditInfo.localTimestamp,
+    authChain: auditInfo.authChain,
+    overwrittenBy: auditInfo.overwrittenBy,
+    isDenylisted: auditInfo.isDenylisted,
+    denylistedContent: auditInfo.denylistedContent
+  }
+  return {
+    status: 200,
+    body
+  }
+}
